ARM: shmobile: R-Car: Shrink rcar_sysc_ch size
Shrink the individual fields in struct rcar_sysc_ch, as unsigned long or
int is overkill:
  - chan_offs contains a register offset relative to a base value
    (< 512),
  - chan_bit and isr_bit contain bit indices (0-31).

This reduces the size of each instance from 3 (4 on 64-bit) 32-bit words
to 1 32-bit word.

+++ arch/arm/mach-shmobile/pm-rcar.h
@@ -3,9 +3,9 @@ FASTVC-BENCH-CORPUS:linux-main-100k-v1-e0bd41dc-8e12-4f1b-978d-a3645ae59da0
 #define PM_RCAR_H
 
 struct rcar_sysc_ch {
-	unsigned long chan_offs;
-	unsigned int chan_bit;
-	unsigned int isr_bit;
+	u16 chan_offs;
+	u8 chan_bit;
+	u8 isr_bit;
 };
 
 int rcar_sysc_power_down(struct rcar_sysc_ch *sysc_ch);
